Fix linker script alignment of TLS sections
The linker script syntax means something different if ALIGN(...) is before
or after the : in an input section line.

With it after the : it sets the sh_addralign of the output section. This is
what's necessary for the resolution of tpoff(var) to take the alignment
into account correctly.

We need the sh_addralign to be only 16, not 32.

BUG= http://code.google.com/p/nativeclient/issues/detail?id=2330
TEST= hand-tested